# docs-site/schemas/guest-share.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Static export of guest share / remote MCP HTTP paths from the host GUI app. Not served on the public relay. Alpha \u2014 may change without notice. Document commands: schemas/document-commands.schema.json",
    "title": "Podcast MCP guest share HTTP",
    "version": "alpha"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/rec/{token}/bootstrap": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/rec/{token}/features": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/rec/{token}/upload": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "description": "keeper or room_tone",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "keeper",
                    "room_tone"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "keeper or room_tone",
              "title": "Kind"
            }
          },
          {
            "in": "header",
            "name": "X-Record-Participant",
            "required": true,
            "schema": {
              "title": "X-Record-Participant",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Record-Lease",
            "required": true,
            "schema": {
              "title": "X-Record-Lease",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Record-Participant",
            "required": true,
            "schema": {
              "title": "X-Record-Participant",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Record-Lease",
            "required": true,
            "schema": {
              "title": "X-Record-Lease",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "take_index",
            "required": true,
            "schema": {
              "title": "Take Index",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "segment_index",
            "required": true,
            "schema": {
              "title": "Segment Index",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "part_seq",
            "required": true,
            "schema": {
              "title": "Part Seq",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sha256",
            "required": false,
            "schema": {
              "default": "",
              "title": "Sha256",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "file_sha256",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "File Sha256"
            }
          },
          {
            "in": "query",
            "name": "final",
            "required": false,
            "schema": {
              "default": false,
              "title": "Final",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "join_offset_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Join Offset Ms"
            }
          },
          {
            "description": "keeper or room_tone",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "keeper",
                    "room_tone"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "keeper or room_tone",
              "title": "Kind"
            }
          },
          {
            "in": "header",
            "name": "X-Record-Participant",
            "required": true,
            "schema": {
              "title": "X-Record-Participant",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Record-Lease",
            "required": true,
            "schema": {
              "title": "X-Record-Lease",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
...
  }
}